requirements_test.txt
pytest
pytest-asyncio
pytest-homeassistant-custom-component
vi_api_client @ git+https://github.com/ignazhabibi/vi_api_client.git
